revisi #3

+++ lib/models/cart.dart
@@ -3,7 +3,7 @@ class Cart {
   int bookId;
   int cartQty;
   String bookTitle;
-  int bookPrice; 
+  double bookPrice; // Perbaikan disini, menggunakan tipe double
 
   Cart({
     required this.cartId,
@@ -16,10 +16,10 @@ class Cart {
   factory Cart.fromJson(Map<String, dynamic> json) {
     return Cart(
       cartId: json['cart_id'],
-      bookId:int.tryParse(json['book_id'].toString()) ?? 0,
+      bookId: int.tryParse(json['book_id'].toString()) ?? 0,
       cartQty: int.tryParse(json['cart_qty'].toString()) ?? 0,
       bookTitle: json['book_title'],
-      bookPrice: int.tryParse(json['book_price'].toString()) ?? 0,
+      bookPrice: double.tryParse(json['book_price'].toString()) ?? 0.0, // Perbaikan disini
     );
   }
 }
